Converted entire bot to hashmap using exports function. Converted all switch cases into variable methods. All works with no seen errors so far.
